requirements.txt

fastapi==0.95.2
uvicorn==0.22.0
pytesseract==0.3.10
opencv-python==4.7.0.72
pillow==9.5.0
nest_asyncio==1.5.6
